+++ internal/transport/http/config.go
@@ -1,4 +1,4 @@
-package transport
+package http
 
 import "os"
