+++ internal/transport/http/handler/get_original_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/zenoleg/shortener/internal/domain"
 	"github.com/zenoleg/shortener/internal/storage"
-	"github.com/zenoleg/shortener/internal/transport"
+	http2 "github.com/zenoleg/shortener/internal/transport/http"
 	"github.com/zenoleg/shortener/internal/transport/http/handler/mocks"
 )
 
@@ -85,7 +85,7 @@ func TestGetOriginalURLHandler_Handle(t *testing.T) {
 }
 func makeGetOriginalURLTestEnv(t *testing.T, handler GetOriginalURLHandler) *httpexpect.Expect {
 	e := echo.New()
-	e.Binder = transport.NewValidationBinder()
+	e.Binder = http2.NewValidationBinder()
 	e.HideBanner = true
 
 	e.Use(middleware.Recover())
