+++ internal/transport/http/http.go
@@ -1,4 +1,4 @@
-package transport
+package http
 
 import (
 	"context"
