--- internal/transport/request.go
@@ -1,35 +1,0 @@
-package transport
-
-import validation "github.com/go-ozzo/ozzo-validation/v4"
-
-type (
-	ShortenRequest struct {
-		URL string `json:"url" query:"url"`
-	}
-
-	OriginalRequest struct {
-		URL string `query:"url"`
-	}
-
-	RedirectRequest struct {
-		ShortID string `param:"shortID"`
-	}
-)
-
-func (r ShortenRequest) Validate() error {
-	return validation.ValidateStruct(&r,
-		validation.Field(&r.URL, validation.Required),
-	)
-}
-
-func (r OriginalRequest) Validate() error {
-	return validation.ValidateStruct(&r,
-		validation.Field(&r.URL, validation.Required),
-	)
-}
-
-func (r RedirectRequest) Validate() error {
-	return validation.ValidateStruct(&r,
-		validation.Field(&r.ShortID, validation.Required),
-	)
-}
